+++ demos/employees/src/main/java/is/codion/demos/employees/ui/EmployeesAppPanel.java
@@ -63,7 +63,8 @@ public class EmployeesAppPanel extends EntityApplicationPanel<EmployeesAppModel>
 		EntityPanel departmentPanel = new EntityPanel(departmentModel,
 						new DepartmentEditPanel(departmentModel.editModel()),
 						new DepartmentTablePanel(departmentModel.tableModel()),
-						config -> config.detailLayout(entityPanel -> TabbedDetailLayout.builder(entityPanel)
+						config -> config.detailLayout(entityPanel -> TabbedDetailLayout.builder()
+										.panel(entityPanel)
 										.splitPaneResizeWeight(0.4)
 										.build()));
 		departmentPanel.detailPanels().add(employeePanel);